backend/pyproject.toml
[tool.poetry]
name = "rekindled"
version = "0.1.0"
description = ""
authors = ["s7m4b4 <8148044+s7m4b4@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"
factory-boy = "^3.2.1"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"
django-filter = "21.1"
hashids = "^1.3.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
